[FLINK-8178][network] Introduce not threadsafe write only BufferBuilder
While Buffer class is used in multithreaded context it requires synchronisation.
Previously it was miss-leading and unclear, suggesting that RecordSerializer should
take into account synchronisation of the Buffer that's holding. With NotThreadSafe
BufferBuilder there is now clear separation between single-threaded writing/creating
a BufferBuilder and multithreaded Buffer handling/retaining/recycling.

This increases throughput of network stack by factor of 2, because previously
method getMemorySegment() was called twice per record and it is a synchronized
method on recycleLock, while RecordSerializer is sole owner of the Buffer at
this point, so synchronisation is not needed.